API

Blueprints.Blueprint — Type

Blueprint

A blueprint defined using B. Its positional and keyword arguments can be retrieved using getindex. The function can be accessed through its func field.

Examples

julia> blueprint = B(repeat, [1,2]; inner=100);
julia> blueprint.func
repeat
julia> blueprint[1]
[1,2]
julia> blueprint[:inner]
100

See also: B, construct.

Blueprints.CachedBlueprint — Type

CachedBlueprint

A cached blueprint, defined using CachedB.

Fields

filename: the JLD2 file used as cache

groupname: the path of the object within the file.

blueprint: the blueprint that is to be cached

Blueprints.MapPolicy — Type

MapPolicy(map_func [; maxconcurrency=nothing])

Simple parallel execution policy that allows specifying an implementation of parallel map. In this way, it is possible to construct blueprints either using threads or distributed workflows.

maxconcurrency allows setting a maximum number items that map_func is called on at once. Use this to limit memory consumption or encourage early caching. By default, there is no limit.

Blueprints.PhonyBlueprint — Type

Blueprints.PhonyBlueprint(constructor, dependencies, blueprint::Blueprint)

Looks and serializes like blueprint, but actually executes constructor(dependencies).

This is useful mostly in cases where your calculation serializes in a very inefficient way or contains closures. In those cases, you can (at your own risk) mask the calculation behind a stand-in blueprint so that the result is still pure in terms of the arguments and parameters of blueprint.

Blueprints.B — Method

B(func, args...; params...)

Defines a blueprint for the evaluation of func(args...; params...).

Note

It is important that the blueprinted function is pure: The function must not modify its inputs and its result must not depend on side effects. This is because construct will automatically apply memoization to avoid building the same blueprint more than once. Here, same is defined by isequal.

Blueprints.CachedB — Method

CachedB(filename, func, args...; params...)
CachedB((filename, groupname), func, args...; params...)

Defines a cached blueprint for the evaluation of func(args...; params...).

On construct, the result is written to the JLD2 file called filename under the groupname groupname. If the cache already exists, it is loaded instead.

If groupname is omitted, a default group name is chosen using default_groupname.

Note

Using the default group name only works if all constituents of the blueprint have a repr that remains stable between Julia sessions. This is not true for closures.

Blueprints.CachedB — Method

CachedB(filename, blueprint)
CachedB((filename, groupname), blueprint)

Promotes a regular blueprint into a cached blueprint.

Blueprints.construct — Method

construct(x; policy=MapPolicy(map), copy=true, readonly=false)

If x is a blueprint, constructs it.

If x is a cached blueprint, first tries to load from cache, otherwise writes the cache after construction.

If x implements Blueprints.dependencies, constructs its dependencies recursively and then x.

Else, returns x.

Each blueprint will only be constructed once. If another one which is identical (===) to it appears, the result is memoized.

policy can be used to set an execution policy for parallelization of independent calculations. At the moment, only MapPolicy is supported, which allows you to specify a parallel map implementation. In the future, additional policies may be implemented.

If copy is enabled, intermediate results will always be deepcopied before they are passed to constructors. This avoids unexpected data dependencies due to memoization, in cases like

a = B(zeros,4)
result = construct([a,a]; copy=false) # danger!
result[1][1] = 1
result

# output
2-element Vector{Any}:
 [1.0, 0.0, 0.0, 0.0]
 [1.0, 0.0, 0.0, 0.0]

If readonly is set, construction of cached blueprints will fail if their caches do not yet exist. In this mode, construct can be called from multiple processes without coordination on cached blueprints. ```

Blueprints.default_groupname — Method

Blueprints.default_groupname(x) -> String

The default groupname that is used by CachedB if only a filename is given. This will usually be a textual representation of the function call. If the representation is longer than Blueprints.MAX_CACHE_GROUPNAME_LENGTH, it is replaced by its hash instead.

Blueprints.dependencies — Function

Blueprints.dependencies(x) -> (AbstractVector, Function)

Returns the dependencies of x as a tuple of a vector deps and a constructor f so that f(deps) == x.

Defining this method for a custom type allows to construct x transparently: If any of the dependencies are Blueprints, they will be recursively constructed before the whole type is. For example, constructing an array of blueprints will return an array of constructed blueprints.

Blueprints.is_cached — Method

is_cached(bp::CachedBlueprint)

Returns true if the cache of bp is populated.